cMusicManager

+++ src/audio/DMAudio.h
@@ -253,7 +253,7 @@ public:
 	bool IsMissionAudioSampleFinished(void);
 	void ClearMissionAudio(void);
 
-	int32 GetRadioInCar(void);
+	uint8 GetRadioInCar(void);
 	void SetRadioInCar(uint32 radio);
 	void SetRadioChannel(int8 radio, int32 pos);
 };
